Last changes
It is now checked if the needed toolboxes (Image Processing, Statistics and Machine learning) are installed, as they are needed to run the program.
Small changes to the code, removal of properties /functions not used anymore. Addition of the app package as "standalone" variant.

+++ SaveStack.m
@@ -19,9 +19,9 @@ stackInfo.Height  = size(stack,1);
 stackInfo.nFrames = size(stack,3);
 
 stackInfo.mean   = mean(stack,3);
-stackInfo.median = median(stack,3);
+stackInfo.median = median(double(stack),3);
 %std only works on single/double
-stackInfo.std    = std(single(stack),0,3);
+stackInfo.std    = std(double(stack),0,3);
 
 % The color bar should be adjusted to the first image, we still
 % take the first five, in case the first one is artefact-ridden
